gdk: Convert mouse position to doubles, add new getters
We've long had double precision mouse coordinates on wayland (e.g.
when rotating a window) but with the new scaling we even have it on
X (and, its also in Xinput2), so convert all the internal mouse/device
position getters to use doubles and add new accessors for the
public APIs that take doubles instead of ints.

+++ gdk/broadway/gdkwindow-broadway.c
@@ -750,8 +750,8 @@ gdk_broadway_window_get_frame_extents (GdkWindow    *window,
 static gboolean
 gdk_window_broadway_get_device_state (GdkWindow       *window,
 				      GdkDevice       *device,
-				      gint            *x,
-				      gint            *y,
+				      gdouble         *x,
+				      gdouble         *y,
 				      GdkModifierType *mask)
 {
   GdkWindow *child;
