+++ gdk/x11/gdkwindow-x11.c
@@ -3240,8 +3240,8 @@ gdk_x11_window_get_frame_extents (GdkWindow    *window,
 static gboolean
 gdk_window_x11_get_device_state (GdkWindow       *window,
                                  GdkDevice       *device,
-                                 gint            *x,
-                                 gint            *y,
+                                 gdouble         *x,
+                                 gdouble         *y,
                                  GdkModifierType *mask)
 {
   GdkWindow *child;
